feat: Created useProportions composable.
Created useProportions composable for setting up main and overlay canvas proportions based on document width..
